+++ Helpers/Utilities/Price.cs
@@ -3,8 +3,8 @@ namespace PEAS.Helpers.Utilities
 {
     public class Price
     {
-        public const long FreePrice = 0;
-        public const long MinPrice = 10_00;
-        public const long MaxPrice = 5000_00;
+        public const int FreePrice = 0;
+        public const int MinPrice = 10_00;
+        public const int MaxPrice = 5000_00;
     }
 }
